Spær distance rettet til i SVG

+++ src/main/java/dat/startcode/model/services/SVG.java
@@ -16,9 +16,9 @@ public class SVG {
             "y=\"%d\"   " +
             " preserveAspectRatio=\"xMinYMin\">";
 
-    private final String rectTemplate = "<rect x=\"%d\" y=\"%d\" height=\"%d\" width=\"%d\" style=\"stroke:#000000; fill: #ffffff\" />";
+    private final String rectTemplate = "<rect x=\"%f\" y=\"%d\" height=\"%d\" width=\"%f\" style=\"stroke:#000000; fill: #ffffff\" />";
 
-    private final String lineTemplate = "<line x1=\"%d\" y1=\"%d\" x2=\"%d\" y2=\"%d\" style=\"stroke:#000000; fill: #ffffff\" />";
+    private final String lineTemplate = "<line x1=\"%d\" y1=\"%d\" x2=\"%f\" y2=\"%d\" style=\"stroke:#000000; fill: #ffffff\" />";
 
 
 
@@ -31,11 +31,11 @@ public class SVG {
         svg.append(String.format(headerTemplate, height, width, viewBox, x, y));
     }
 
-    public void addRect(int x, int y, int height, int width) {
+    public void addRect(float x, int y, int height, float width) {
         svg.append(String.format(rectTemplate, x, y, height, width));
     }
 
-    public void addLine(int x1, int y1, int x2, int y2) {
+    public void addLine(int x1, int y1, float x2, int y2) {
         svg.append(String.format(lineTemplate, x1, y1, x2, y2));
     }
